+++ R/groups.R
@@ -215,25 +215,6 @@ group_model.mhmm <- function(x, type = "relative", scaling = character(0L),
   )
 }
 
-# #' @export
-# #' @rdname group_model
-# group_model.tna_mmm <- function(x, type = "relative", scaling = character(0L),
-#                                 groupwise = FALSE, params = list(),
-#                                 na.rm = TRUE, ...) {
-#   check_missing(x)
-#   check_class(x, "tna_mmm")
-#   group_model.default(
-#     x = x$data,
-#     group = x$assignments,
-#     type = type,
-#     scaling = scaling,
-#     groupwise = groupwise,
-#     params = params,
-#     na.rm = na.rm,
-#     ...
-#   )
-# }
-
 #' @export
 #' @rdname group_model
 group_model.tna_clustering <- function(x, type = "relative",